11-07 додав ImageModifyAbstract (виніс загальний функціонал) змінив GDIImageModifyFile -> GDIImageModify (залишив функціонал GDImage) видалив SaveFile
